Crate qrcode_static

Functions

png_qr

Function to generate static qr code from Vec

png_qr_from_hex

Function to generate static qr code from hex string

png_qr_from_string

Historically was used to generate static qr code from hex string for example in signatures. Used for strings.